Make sure that the PUT/POST actually needs to occur
Otherwise, every "Save" of the context sends a POST for every object we have added locally and a PUT for everything modified, regardless of whether it is actually a new/updated object from the server's perspective.

By checking the body of our PUT we know if anything has actually changed, and by checking whether the object has an "id" we can tell if it came from the server or was created locally.